+++ src/histogram/list_pool.rs
@@ -159,7 +159,7 @@ impl ListPool {
     fn free(&mut self, block: usize, sclass: SizeClass) {
         let sclass = sclass as usize;
         // Insert the block on the free list which is a single linked list.
-        self.data[block] = self.free[sclass] as u32;
+        self.data[block] = self.free[sclass];
         self.free[sclass] = block as u32
     }
